exercice1/requirements_api.txt
fastapi
uvicorn
elasticsearch<7.14.0
black
python-dotenv
requests